--- crates/profile/models/persona/src/persona/mod.rs
@@ -1,3 +1,0 @@
-mod persona;
-
-pub use persona::*;
