+++ crates/sargon_SPLIT_ME/src/sargon_os/transactions/support/signable_manifest_summary.rs
