--- crates/sargon_SPLIT_ME/src/system/mod.rs
@@ -1,10 +1,0 @@
-mod bios;
-mod interactors;
-mod sargon_os;
-mod subsystems;
-mod test_diagnose_instance_duplicates;
-
-pub use bios::*;
-pub use interactors::*;
-pub use sargon_os::*;
-pub use subsystems::*;
